# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ShopPro 智能SCRM系统 API
  version: 1.0.0
  description: ShopPro 智能客户关系管理系统 RESTful API 规范 - 提供完整的客户管理、销售跟进、数据分析等功能接口
servers:
  - url: http://localhost:8080/api
    description: 本地开发
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, minimum: 1, default: 1, example: 1 }
      description: 页码，最小为 1，默认 1
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10, example: 10 }
      description: 每页大小，范围 [1,100]，默认 10
...
  schemas:
    ApiResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data: {}
...
    PageKnowledgeResult:
      type: object
      properties:
        records:
          type: array
          items: { $ref: '#/components/schemas/Knowledge' }
        current: { type: integer }
        size: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    Knowledge:
      type: object
      properties:
        id: { type: integer, format: int64 }
        categoryId: { type: integer, format: int64 }
        title: { type: string }
        content: { type: string }
        tags: { type: string, nullable: true }
        viewCount: { type: integer, nullable: true }
        likeCount: { type: integer, nullable: true }
        status: { type: integer }
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: OK }
...
  /knowledge/page:
    get:
      tags: [Knowledge]
      summary: 分页查询知识
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageKnowledgeResult'
...
  /knowledge:
    put:
      tags: [Knowledge]
      summary: 更新知识条目
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Knowledge' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Knowledge' }
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /knowledge/{id}:
    get:
      tags: [Knowledge]
      summary: 获取知识详情
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Knowledge' }
        '404':
          description: 未找到该知识
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /knowledge/category/{categoryId}:
    get:
      tags: [Knowledge]
      summary: 根据分类获取知识列表
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Knowledge' }
...